Move LocationPlan cache objects into Grade
This is a major rewrite that moves a lot of the data that was previously
in LocationPlan into Grade

Specifically, everything needed to analyze and grade the current
situation is now in centralized in Grade.  The only things left in
LocationPlan are: chosen_grade, WiW state, Approval State.

This opens the door for many upcoming changes with regard to modifying
and updating visit/volume data periodically as a Schedule month
approaches.